Added XDG desktop entry for use on Linux/Unix
Adjusted install rules to install both entry and icon in appropriate
location.  The data/e-foto.png file was generated from efoto.ico using
'convert efoto.ico data/e-foto.png'.  It would be better if it had at least
64x64 pixel resolution or was a vector image in SVG format.

The naming of this tool is  inconsistent, ie efoto versus e-foto.  I used
the reverse DNS notation with efoto for the file name, and e-foto for the icon,
but suspect it would be better if one consistent naming standard was used.